refactor: inline calcMarquee and remove tool flags
