+++ src/utils/index.js
@@ -13,36 +13,6 @@ export function getPixelUnion(props = []) {
     return [...set].map(keyToCoord);
 }
 
-export function calcMarquee(startEvent, currentEvent, viewportStore, viewportEl) {
-    if (!startEvent || !currentEvent || !viewportEl) return { visible: false, x: 0, y: 0, w: 0, h: 0 };
-    const rect = viewportEl.getBoundingClientRect();
-    const style = getComputedStyle(viewportEl);
-    const leftBase = rect.left + parseFloat(style.paddingLeft) + viewportStore.stage.offset.x;
-    const topBase = rect.top + parseFloat(style.paddingTop) + viewportStore.stage.offset.y;
-    const scale = viewportStore.stage.scale;
-    const width = viewportStore.stage.width;
-    const height = viewportStore.stage.height;
-    const left = Math.min(startEvent.clientX, currentEvent.clientX) - leftBase;
-    const top = Math.min(startEvent.clientY, currentEvent.clientY) - topBase;
-    const right = Math.max(startEvent.clientX, currentEvent.clientX) - leftBase;
-    const bottom = Math.max(startEvent.clientY, currentEvent.clientY) - topBase;
-    const minX = Math.floor(left / scale),
-        maxX = Math.floor((right - 1) / scale);
-    const minY = Math.floor(top / scale),
-        maxY = Math.floor((bottom - 1) / scale);
-    const minx = clamp(minX, 0, width - 1),
-        maxx = clamp(maxX, 0, width - 1);
-    const miny = clamp(minY, 0, height - 1),
-        maxy = clamp(maxY, 0, height - 1);
-    return {
-        visible: true,
-        x: minx,
-        y: miny,
-        w: (maxx >= minx) ? (maxx - minx + 1) : 0,
-        h: (maxy >= miny) ? (maxy - miny + 1) : 0,
-    };
-}
-
 export function ensureCheckerboardPattern(target = document.body) {
     const { PATTERN_ID, COLOR_A, COLOR_B, REPEAT } = CHECKERBOARD_CONFIG;
     const id = PATTERN_ID;
